+++ buildChecks.js
@@ -2,10 +2,10 @@
 const fs = require('fs')
 const cp = require('child_process')
 
-function checkIfPrebuildExists() {
+function checkIfPrebuildExists () {
   try {
     const bindings = require('./binding')
-    if (!bindings) throw 'Bindings are undefined'
+    if (!bindings) throw Error('Bindings are undefined')
     console.log('[leveldb] not building as already have prebuild')
     return true
   } catch (e) {
@@ -25,18 +25,16 @@ if (process.env.SKIP_BUILD) {
   runCmake = false
 }
 
-async function runChecks() {
+async function runChecks () {
   if (!fs.existsSync('./leveldb-mcpe/include')) {
-
     console.info('Cloning submodules...')
     cp.execSync('git submodule init', { stdio: 'inherit' })
     cp.execSync('git submodule update', { stdio: 'inherit' })
 
     if (!fs.existsSync('./leveldb-mcpe/include')) { // npm install does not clone submodules...
-
       cp.execSync('git clone https://github.com/extremeheat/leveldb-mcpe') // so do it manually
 
-      if (!fs.existsSync('./leveldb-mcpe/include')) { //gie up
+      if (!fs.existsSync('./leveldb-mcpe/include')) { // gie up
         console.error('******************* READ ME ****************\n')
         console.error(' Failed to install git submodules. Please create an issue at https://github.com/extremeheat/node-leveldb-zlib\n')
         console.error('******************* READ ME ****************\n')
@@ -45,10 +43,9 @@ async function runChecks() {
     }
   }
 
-  if (process.platform == 'win32') {
+  if (process.platform === 'win32') {
     if (!process.env.CMAKE_TOOLCHAIN_FILE) {
-
-      const exec = require("child_process");
+      const exec = require('child_process')
       // Try to set CMAKE_TOOLCHAIN_FILE with pre-packaged vcpkg
       exec.execSync('cd helpers && win-build.bat')
 
@@ -61,15 +58,15 @@ async function runChecks() {
         console.log('Using pre-bundled vcpkg')
       }
     }
-  } else if (process.platform == 'darwin') {
-    if (!fs.existsSync(`/usr/local/opt/zlib/include/`)) {
+  } else if (process.platform === 'darwin') {
+    if (!fs.existsSync('/usr/local/opt/zlib/include/')) {
       console.error('******************* READ ME ****************\n')
       console.error(' zlib was not found. Run `xcode-select --install` and try again.\n')
       console.error(' The build below probably failed.\n')
       console.error('******************* READ ME ****************\n')
     }
   } else {
-    if (!fs.existsSync(`/usr/include/zlib.h`) && !fs.existsSync(`/usr/local/include/zlib.h`)) {
+    if (!fs.existsSync('/usr/include/zlib.h') && !fs.existsSync('/usr/local/include/zlib.h')) {
       console.error('******************* READ ME ****************\n')
       console.error(' zlib headers were not found. If the build fails, try `sudo apt-get install libz-dev`\n')
       console.error('******************* READ ME ****************\n')
@@ -81,7 +78,7 @@ if (runCmake) {
   runChecks().then(
     () => {
       console.log('Build checks are passing! Building...')
-      cp.execSync(`cmake-js compile`, {stdio: 'inherit'})
+      cp.execSync('cmake-js compile', { stdio: 'inherit' })
     }
   )
 }