+++ helpers/buildPath.js
@@ -1,25 +1,23 @@
-const fs = require('fs')
 const os = require('os')
-const path = require('path')
 
 module.exports = {
-  getPath() {
-    let _osVersion = os.release()
+  getPath () {
+    const _osVersion = os.release()
 
-    let plat = process.platform
-    let arch = process.arch
-    let ver = _osVersion.split('.', 1)
+    const plat = process.platform
+    const arch = process.arch
+    const ver = _osVersion.split('.', 1)
 
-    let bpath = `./prebuilds/${plat}-${ver}-${arch}/`
+    const bpath = `./prebuilds/${plat}-${ver}-${arch}/`
     return bpath
   },
-  
-  getPlatformString() {
-    let _osVersion = os.release()
 
-    let plat = process.platform
-    let arch = process.arch
-    let ver = _osVersion.split('.', 1)
+  getPlatformString () {
+    const _osVersion = os.release()
+
+    const plat = process.platform
+    const arch = process.arch
+    const ver = _osVersion.split('.', 1)
     return `${plat}-${ver}-${arch}`
   }
 }