+++ HISTORY.md
@@ -1,3 +1,10 @@
+## 1.1.0
+* Fix `LevelDB.repair()` and `LevelDB.destory()` methods, and make them static, use as such:
+```js
+import { LevelDB } from 'leveldb-zlib'
+await LeveDB.repair('./db')
+```
+
 ## 1.0
 * All LevelDB APIs now use Promise instead of callbacks
 * Improve documentation